Move ConnectionConfig to UtpConfig
