+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod cid;
+pub mod config;
 pub mod congestion;
 pub mod conn;
 pub mod event;
