+++ tests/socket.rs
@@ -2,24 +2,24 @@ use std::net::SocketAddr;
 use std::sync::Arc;
 
 use utp_rs::cid;
-use utp_rs::conn::ConnectionConfig;
+use utp_rs::config::UtpConfig;
 use utp_rs::socket::UtpSocket;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn socket() {
     tracing_subscriber::fmt::init();
 
-    let conn_config = ConnectionConfig::default();
+    let conn_config = UtpConfig::default();
 
     let data_one = vec![0xef; 8192 * 2 * 2];
     let data_one_recv = data_one.clone();
 
     let recv_addr = SocketAddr::from(([127, 0, 0, 1], 3400));
-    let recv = UtpSocket::bind(recv_addr).await.unwrap();
+    let recv = UtpSocket::bind(recv_addr, conn_config).await.unwrap();
     let recv = Arc::new(recv);
 
     let send_addr = SocketAddr::from(([127, 0, 0, 1], 3401));
-    let send = UtpSocket::bind(send_addr).await.unwrap();
+    let send = UtpSocket::bind(send_addr, conn_config).await.unwrap();
     let send = Arc::new(send);
 
     let recv_one_cid = cid::ConnectionId {
@@ -35,10 +35,7 @@ async fn socket() {
 
     let recv_arc = Arc::clone(&recv);
     let recv_one_handle = tokio::spawn(async move {
-        let mut stream = recv_arc
-            .accept_with_cid(recv_one_cid, conn_config)
-            .await
-            .unwrap();
+        let mut stream = recv_arc.accept_with_cid(recv_one_cid).await.unwrap();
         let mut buf = vec![];
         let n = stream.read_to_eof(&mut buf).await.unwrap();
         tracing::info!(cid.send = %recv_one_cid.send, cid.recv = %recv_one_cid.recv, "read {n} bytes from uTP stream");
@@ -49,10 +46,7 @@ async fn socket() {
 
     let send_arc = Arc::clone(&send);
     tokio::spawn(async move {
-        let mut stream = send_arc
-            .connect_with_cid(send_one_cid, conn_config)
-            .await
-            .unwrap();
+        let mut stream = send_arc.connect_with_cid(send_one_cid).await.unwrap();
         let n = stream.write(&data_one).await.unwrap();
         assert_eq!(n, data_one.len());
 
@@ -74,10 +68,7 @@ async fn socket() {
     };
 
     let recv_two_handle = tokio::spawn(async move {
-        let mut stream = recv
-            .accept_with_cid(recv_two_cid, conn_config)
-            .await
-            .unwrap();
+        let mut stream = recv.accept_with_cid(recv_two_cid).await.unwrap();
         let mut buf = vec![];
         let n = stream.read_to_eof(&mut buf).await.unwrap();
         tracing::info!(cid.send = %recv_two_cid.send, cid.recv = %recv_two_cid.recv, "read {n} bytes from uTP stream");
@@ -87,10 +78,7 @@ async fn socket() {
     });
 
     tokio::spawn(async move {
-        let mut stream = send
-            .connect_with_cid(send_two_cid, conn_config)
-            .await
-            .unwrap();
+        let mut stream = send.connect_with_cid(send_two_cid).await.unwrap();
         let n = stream.write(&data_two).await.unwrap();
         assert_eq!(n, data_two.len());
